Develop

+++ XUnitBlog.Domain/Repositories/IUserRepository.cs
@@ -7,7 +7,7 @@ public interface IUserRepository
     Task<User?> GetUserByUserName(string userName);
     Task<User?> GetUserByEmail(string email);
     Task AddAsync(User user);
-    Task<User?> GetUserById(int userId);
+    Task<User?> GetUserById(long userId);
     Task<IList<User>> GetAll();
     Task UpdateAsync(User user);
 }
